# gateway/src/test/resources/orders-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "Generated server url"
        }
    ],
    "tags": [
        {
            "name": "orders",
            "description": "Create, inspect and pay for orders"
        },
        {
            "name": "catalog",
            "description": "Read-only product catalogue"
        }
    ],
    "paths": {
        "/api/orders": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "List orders, optionally filtered by customer or status",
                "operationId": "listOrders",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "PENDING",
                                "PAID",
                                "CANCELLED"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Create an order",
                "description": "Mutating. Send an Idempotency-Key so a retried call does not create a second order.",
                "operationId": "createOrder",
                "parameters": [
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "description": "Replay-safe key, echoed by the caller on retry",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{id}/payments": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "List payments captured against an order",
                "operationId": "listPayments",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PaymentResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Capture payment for an order",
                "description": "Mutating and money-moving. Requires an Idempotency-Key in production profiles.",
                "operationId": "capturePayment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CapturePaymentRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaymentResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Get one order by id",
                "operationId": "getOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/catalog/items": {
            "get": {
                "tags": [
                    "catalog"
                ],
                "summary": "List catalogue items",
                "description": "Read-only. Safe for a guest key.",
                "operationId": "listCatalogItems",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CatalogItem"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateOrderRequest": {
                "type": "object",
                "properties": {
                    "customerId": {
                        "type": "string",
                        "minLength": 1
                    },
                    "sku": {
                        "type": "string",
                        "minLength": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "format": "int32",
                        "maximum": 1000,
                        "minimum": 1
                    }
                },
                "required": [
                    "customerId",
                    "sku"
                ]
            },
            "OrderResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "customerId": {
                        "type": "string"
                    },
                    "sku": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "amountCents": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CapturePaymentRequest": {
                "type": "object",
                "properties": {
                    "amountCents": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 1
                    }
                }
            },
            "PaymentResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "orderId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "amountCents": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "status": {
                        "type": "string"
                    },
                    "capturedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "CatalogItem": {
                "type": "object",
                "properties": {
                    "sku": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "unitPriceCents": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "stock": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            }
        }
    }
}